# api/openapi.yaml
openapi: 3.1.0
info:
  title: MESGuard API
  version: 0.1.0-p7-admin-recovery
  description: Current implemented backend contract for authentication, ERP external cases, asynchronous diagnosis tasks, resumable TaskEvent streaming, formal report reading, administrator recovery, and report feedback.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://127.0.0.1:9090
paths:
...
  /api/v1/diagnosis-tasks:
    post:
      operationId: createDiagnosisTask
      summary: 创建异步诊断任务
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisTaskCreateRequest"
      responses:
        "202":
          description: 任务已持久化，等待后续 Outbox Relay/Worker
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosisTaskCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /api/v1/diagnosis-tasks/{taskId}:
    get:
      operationId: getDiagnosisTask
      summary: 查询诊断任务安全摘要
      security:
        - sessionCookie: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 任务状态、快照身份和报告可用性
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosisTaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/diagnosis-tasks/{taskId}/events:
    get:
      operationId: listDiagnosisTaskEvents
      summary: 查询或持续订阅结构化任务事件
      description: application/json 返回游标页；text/event-stream 先从 PostgreSQL 补读，再轮询新事件，终态事件后关闭。断开连接不会取消任务。
      security:
        - sessionCookie: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: afterSeq
          in: query
          description: 排他游标，只返回 seq 大于该值的事件
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: Last-Event-ID
          in: header
          description: SSE 重连游标；提供时优先于 afterSeq
          schema:
            type: string
            pattern: "^[0-9]+$"
        - name: limit
          in: query
          description: 单次 JSON 页或 SSE PostgreSQL 补读批次的最大事件数
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        "200":
          description: 按 seq 升序排列的安全任务事件页或可断点续传的 SSE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosisTaskEventsResponse"
            text/event-stream:
              schema:
                type: string
              example: |-
                retry: 3000

                id: 19
                event: task_started
                data: {"seq":19,"eventType":"task_started","payload":{"attemptCount":1},"payloadSchemaVersion":1,"createdAt":"2026-08-03T06:00:00Z"}
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/diagnosis-tasks/{taskId}/cancel:
    post:
      operationId: cancelDiagnosisTask
      summary: 幂等请求取消诊断任务
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 任务此前已经请求取消或已经取消，未重复写事件
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosisTaskResponse"
        "202":
          description: 已原子写入 cancel_requested 状态和任务事件
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosisTaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mesguard_session
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
        maxLength: 128
  responses:
    BadRequest:
      description: 请求格式或参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未登录或 Session 已失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Origin、CSRF 或权限校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 幂等键冲突或外部工单 fingerprint 已变化
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: 业务输入校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DependencyUnavailable:
      description: PostgreSQL 或 ERP SQL Server 暂时不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    EnvelopeBase:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        requestId:
          type: string
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              additionalProperties: true
...
    DiagnosisTaskCreateRequest:
      type: object
      additionalProperties: false
      required: [externalCaseId, expectedSourceFingerprint, requestText]
      properties:
        externalCaseId:
          type: string
          format: uuid
        expectedSourceFingerprint:
          type: string
          maxLength: 128
        evidenceDataSourceIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        requestText:
          type: string
          minLength: 1
          maxLength: 20000
        requestScope:
          $ref: "#/components/schemas/DiagnosisTaskRequestScope"
        requestScopeSchemaVersion:
          type: integer
          minimum: 1
          default: 1
        attachments:
          type: array
          description: 当前版本暂不接受非空附件列表
          maxItems: 0
          items:
            type: object
        retryOfTaskId:
          type: [string, "null"]
          format: uuid
    DiagnosisTaskRequestScope:
      type: object
      additionalProperties: true
      description: >-
        创建时冻结的任务调查范围。allowedCapabilities 表示本任务允许使用的业务能力，
        不表示对应外部依赖当前健康；未提供时后端按 requestedSkill 推导，仍未提供则默认 [case]。
      properties:
        requestedSkill:
          type: string
          enum: [ticket-diagnosis, code-investigation, sql-investigation]
          default: ticket-diagnosis
        allowedCapabilities:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          default: [case]
          items:
            type: string
            enum: [case, code, sql]
    DiagnosisTaskCreateData:
      type: object
      required: [taskId, status, replayed, createdAt]
      properties:
        taskId:
          type: string
          format: uuid
        status:
          type: string
          const: pending
        replayed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DiagnosisTaskCreateResponse:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/DiagnosisTaskCreateData"
    DiagnosisTask:
      type: object
      required: [taskId, externalCaseId, caseSnapshotId, requestText, requestScope, requestScopeSchemaVersion, status, attemptCount, createdAt, updatedAt, reportAvailable]
      properties:
        taskId: {type: string, format: uuid}
        externalCaseId: {type: string, format: uuid}
        caseSnapshotId: {type: string, format: uuid}
        retryOfTaskId: {type: string, format: uuid}
        requestText: {type: string}
        requestScope:
          $ref: "#/components/schemas/DiagnosisTaskRequestScope"
        requestScopeSchemaVersion: {type: integer, minimum: 1}
        status:
          type: string
          enum: [pending, running, cancel_requested, succeeded, failed, cancelled]
        attemptCount: {type: integer, minimum: 0}
        lastErrorCode: {type: string}
        lastErrorMessage: {type: string}
        startedAt: {type: [string, "null"], format: date-time}
        completedAt: {type: [string, "null"], format: date-time}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        reportAvailable: {type: boolean}
        reportId: {type: string, format: uuid}
    DiagnosisTaskResponse:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/DiagnosisTask"
...
    DiagnosisTaskEvent:
      type: object
      additionalProperties: false
      required: [seq, eventType, payload, payloadSchemaVersion, createdAt]
      properties:
        seq:
          type: integer
          format: int64
          minimum: 1
        eventType:
          type: string
        payload:
          type: object
          additionalProperties: true
        payloadSchemaVersion:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
    DiagnosisTaskEventsData:
      type: object
      additionalProperties: false
      required: [items, afterSeq, nextAfterSeq, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosisTaskEvent"
        afterSeq:
          type: integer
          format: int64
          minimum: 0
        nextAfterSeq:
          type: integer
          format: int64
          minimum: 0
        hasMore:
          type: boolean
    DiagnosisTaskEventsResponse:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/DiagnosisTaskEventsData"
...